Flatten `ext` module
