--- packages/libs/error-stack/src/ext/mod.rs
@@ -1,4 +1,0 @@
-//! Extension traits for foreign types.
-
-pub mod future;
-pub mod result;
